refactor  code

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -8,7 +8,7 @@
  */
 int set_bit(unsigned long int *n, unsigned int index)
 {
-	unsigned long int m;
+	unsigned int m;
 
 	if (index > 63)
 		return (-1);
